+++ integrationtests/src/test/java/io/kroxylicious/proxy/filter/RejectingCreateTopicFilterFactory.java
@@ -6,36 +6,21 @@
 
 package io.kroxylicious.proxy.filter;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
-
 import io.kroxylicious.proxy.plugin.PluginConfigType;
 
-@PluginConfigType(RejectingCreateTopicFilterFactory.RejectingCreateTopicFilterConfig.class)
+@PluginConfigType(RejectingCreateTopicFilter.RejectingCreateTopicFilterConfig.class)
 public class RejectingCreateTopicFilterFactory
-        implements FilterFactory<RejectingCreateTopicFilterFactory.RejectingCreateTopicFilterConfig, RejectingCreateTopicFilterFactory.RejectingCreateTopicFilterConfig> {
+        implements FilterFactory<RejectingCreateTopicFilter.RejectingCreateTopicFilterConfig, RejectingCreateTopicFilter.RejectingCreateTopicFilterConfig> {
 
     @Override
-    public RejectingCreateTopicFilterConfig initialize(FilterFactoryContext context, RejectingCreateTopicFilterConfig config) {
+    public RejectingCreateTopicFilter.RejectingCreateTopicFilterConfig initialize(FilterFactoryContext context, RejectingCreateTopicFilter.RejectingCreateTopicFilterConfig config) {
         // null configuration is allowed, by default null config is invalid
         return config;
     }
 
     @Override
-    public RejectingCreateTopicFilter createFilter(FilterFactoryContext context, RejectingCreateTopicFilterConfig configuration) {
+    public RejectingCreateTopicFilter createFilter(FilterFactoryContext context, RejectingCreateTopicFilter.RejectingCreateTopicFilterConfig configuration) {
         return new RejectingCreateTopicFilter(context, configuration);
     }
 
-    /**
-     * @param withCloseConnection
-     * If true, rejection will also close the connection */
-    public record RejectingCreateTopicFilterConfig(boolean withCloseConnection, ForwardingStyle forwardingStyle) {
-
-        @JsonCreator
-        public RejectingCreateTopicFilterConfig(@JsonProperty(value = "withCloseConnection") boolean withCloseConnection,
-                                                @JsonProperty(value = "forwardingStyle") ForwardingStyle forwardingStyle) {
-            this.withCloseConnection = withCloseConnection;
-            this.forwardingStyle = forwardingStyle == null ? ForwardingStyle.SYNCHRONOUS : forwardingStyle;
-        }
-    }
 }
